requirements.txt
altair==4.1.0
astropy==4.2.1
convertdate==2.3.2
holidays==0.11.3.1
hotstepper==1.0.10
hydralit==1.0.10
hydralit_components==1.0.6
livelossplot==0.5.4
LunarCalendar==0.0.9
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.4
Pillow==8.4.0
prophet==1.0.1
pydeck==0.7.1
python_dateutil==2.8.2
requests==2.25.1
scipy==1.6.2
setuptools==52.0.0.post20210125
spacy_streamlit==1.0.2
srsly==2.4.2
streamlit==1.2.0
sunpy==3.1.1
torch==1.10.0
torch_lr_finder==0.2.1
tqdm==4.59.0
